ClubPost Component화

+++ src/page/clubs/[clubId]/posts/[postId]/ClubPost.tsx
@@ -18,17 +18,14 @@ import useAccessToken from "@/hooks/useAccessToken";
 import { API_ENDPOINT } from "@/App";
 import Textarea from "@/components/Textarea";
 import getPostCategoryWithKey from "@/util/getPostCategoryWithKey";
+import { ModalType } from "@/libs/types";
+import Delete from "@/components/Delete";
+import PostMenuButton from "@/components/PostMenuButton";
+import CommentMenuButton from "@/components/CommentMenuButton";
+import { getDate } from "@/util/getDate";
 
 interface commentFormData {
   comment: string;
-}
-
-const enum ModalType {
-  POST = "post",
-  COMMENT = "comment",
-  DELETE_POST = "delPost",
-  UPDATE_COMMENT = "updateComment",
-  DELETE_COMMENT = "delComment",
 }
 
 let postLikeArr: any[] = [];
@@ -42,11 +39,12 @@ export default function ClubPost() {
   const [like, setLike] = useState<boolean>(false);
   const [inModal, setInModal] = useState<boolean>(false);
   const [modalType, setModalType] = useState<ModalType>();
-  const [selectId, setSelectId] = useState<number>();
-  const [selectComment, setSelectComment] = useState<any[]>([]);
+
+  const [selectedComment, setSelectedComment] = useState<any>(null);
   const {
     state: { post },
   } = useLocation();
+
   const textareaRef = useRef<HTMLTextAreaElement | null>(null);
   const _ = useAutoResizeTextArea(textareaRef);
 
@@ -58,7 +56,7 @@ export default function ClubPost() {
     setValue,
   } = useForm<commentFormData>();
 
-  const { mutate: comment, isLoading: joinLoading } = useMutation(
+  const { mutate: comment, isLoading: commentLoading } = useMutation(
     `boards/${params.postId}/comments`,
     {
       authorized: true,
@@ -70,8 +68,6 @@ export default function ClubPost() {
     token,
   ]);
 
-  useEffect(() => {}, []);
-
   useEffect(() => {
     if (!post) {
       alert("잘못된 접근입니다.");
@@ -79,18 +75,7 @@ export default function ClubPost() {
     }
   }, [post]);
 
-  const postDelete = async () => {
-    const response = await fetch(`${API_ENDPOINT}/boards/${post.id}`, {
-      method: "DELETE",
-      headers: {
-        "Content-Type": "application/json",
-        Authorization: "Bearer " + token,
-      },
-    });
-    console.log(response);
-    navigate(-1);
-  };
-
+  //모달 닫기 기능
   const closeModal = () => {
     setInModal(false);
   };
@@ -106,33 +91,15 @@ export default function ClubPost() {
     }
   };
 
-  const getDate = (e: number) => {
-    let d = new Date(e);
-    let month = d.getMonth() + 1;
-    let date = d.getDate();
-    let hour: number | string = d.getHours();
-    let minute: number | string = d.getMinutes();
-    let amPm = hour >= 12 ? "오후" : "오전";
-    hour =
-      hour > 12
-        ? hour - 12 >= 10
-          ? hour - 12
-          : `0${hour - 12}`
-        : hour >= 10
-        ? hour
-        : `0${hour}`;
-    minute = minute < 10 ? `0${minute}` : minute;
-    return month + "월 " + date + "일 " + amPm + " " + hour + ":" + minute;
-  };
-
+  //댓글 전송 기능
   const onSubmit = async (commentForm: commentFormData) => {
     if (commentForm.comment.length == 0) {
       return;
     }
-    // let replaceComment = commentForm.comment.replace(
-    //   /(?:\r\n|\r|\n)/g,
-    //   "<br/>"
-    // );
+
+    if (commentLoading) {
+      return;
+    }
 
     let item;
     if (userData) {
@@ -146,7 +113,6 @@ export default function ClubPost() {
     }
 
     const response = await comment(item);
-    console.log(response);
 
     if (response.ok) {
       refreshCommentData();
@@ -154,24 +120,23 @@ export default function ClubPost() {
     setValue("comment", "");
   };
 
-  const commentCall = async () => {
-    const response = await fetch(
-      `${API_ENDPOINT}/boards/comments/${selectId}`,
-      {
-        headers: {
-          "Content-Type": "application/json",
-          Authorization: `Bearer ${token}`,
-        },
-      }
-    );
-    const data = await response.json();
-    console.log(data);
-    setSelectComment(data.data);
+  //글 삭제, 댓글 삭제
+  const postDelete = async () => {
+    const response = await fetch(`${API_ENDPOINT}/boards/${post.id}`, {
+      method: "DELETE",
+      headers: {
+        "Content-Type": "application/json",
+        Authorization: "Bearer " + token,
+      },
+    });
+    navigate(-1);
   };
 
   const commentDelete = async () => {
+    if (!selectedComment) return;
+
     const response = await fetch(
-      `${API_ENDPOINT}/boards/comments/${selectId}`,
+      `${API_ENDPOINT}/boards/comments/${selectedComment.id}`,
       {
         method: "DELETE",
         headers: {
@@ -182,30 +147,26 @@ export default function ClubPost() {
     );
 
     if (response.ok) {
-      refreshCommentData();
+      refreshCommentData({
+        ok: true,
+        data: commentData?.data.filter(
+          (comment: any) => comment.id !== selectedComment.id
+        ),
+      });
     }
+    closeModal();
   };
 
-  const { ref } = register("comment");
-  const commentRef = useRef<HTMLTextAreaElement | null>(null);
-
-  const handleResizeHeight = useCallback(() => {
-    if (!commentRef.current) return;
-
-    commentRef.current.style.height = "0";
-    commentRef.current.style.height = commentRef.current?.scrollHeight + "px";
-  }, [commentRef]);
-
-  // 좋아요 기능
-  const getPostLike = localStorage.getItem(
-    `${params.clubId}_${params.postId} like`
-  );
+  // // 좋아요 기능
+  // const getPostLike = localStorage.getItem(
+  //   `${params.clubId}_${params.postId} like`
+  // );
 
   // useEffect(() => {
   //   if (getPostLike !== null) {
   //     postLikeArr = JSON.parse(getPostLike);
   //     const likeUserArr = postLikeArr.filter(
-  //       (item: { id: number }) => item.id == userId
+  //       (item: { id: number }) => item.id == userData?.id
   //     );
 
   //     if (likeUserArr.length !== 0) {
@@ -223,12 +184,9 @@ export default function ClubPost() {
   //     postLikeArr = JSON.parse(getPostLike);
   //     console.log(postLikeArr);
   //     //유저가 이미 좋아요를 눌렀다면?
-  //     if (
-  //       postLikeArr.filter((item: { id: number }) => item.id == userId)
-  //         .length !== 0
-  //     ) {
+  //     if (postLikeArr.some((item: { id: number }) => item.id == userData?.id)) {
   //       postLikeArr = postLikeArr.filter(
-  //         (item: { id: number }) => item.id !== userId
+  //         (item: { id: number }) => item.id !== userData?.id
   //       );
   //       localStorage.setItem(
   //         `${params.clubId}_${params.postId} like`,
@@ -237,7 +195,7 @@ export default function ClubPost() {
   //     }
   //     //좋아요를 누르지 않았다면
   //     else {
-  //       postLikeArr.unshift(user);
+  //       postLikeArr.unshift({ id: userData?.id });
   //       localStorage.setItem(
   //         `${params.clubId}_${params.postId} like`,
   //         JSON.stringify(postLikeArr)
@@ -246,7 +204,7 @@ export default function ClubPost() {
   //   }
   //   // 글에 좋아요가 0일때
   //   else {
-  //     postLikeArr.unshift(user);
+  //     postLikeArr.unshift({ id: userData?.id });
   //     localStorage.setItem(
   //       `${params.clubId}_${params.postId} like`,
   //       JSON.stringify(postLikeArr)
@@ -260,105 +218,33 @@ export default function ClubPost() {
         modalType &&
         {
           post: (
-            <div className=" w-[200px] h-[90px] flex cursor-pointer bg-white text-left flex-col absolute justify-evenly top-2 right-0 border-[1px] border-solid z-[100]">
-              <Link
-                to={`/clubs/${params.clubId}/update_post/${params.postId}`}
-                state={post}
-              >
-                <div className="h-[40px] p-3 text-[16px]">게시글 수정</div>
-              </Link>
-              <div
-                className="h-[40px] p-3 text-[16px]"
-                onClick={() => setModalType(ModalType.DELETE_POST)}
-              >
-                게시글 삭제
-              </div>
-            </div>
-          ),
-          comment: (
-            <Overlay onClick={closeModal}>
-              <div
-                onClick={(e) => e.stopPropagation()}
-                className=" w-full h-[100px] mt-auto mb-auto ml-3 mr-3 flex bg-white self-end flex-col"
-              >
-                <div
-                  className="h-[50px] p-4 text-[16px]"
-                  onClick={() => {
-                    commentCall();
-                    setModalType(ModalType.UPDATE_COMMENT);
-                  }}
-                >
-                  수정하기
-                </div>
-                <div
-                  className="h-[50px] p-4 text-[16px]"
-                  onClick={() => {
-                    setModalType(ModalType.DELETE_COMMENT);
-                  }}
-                >
-                  삭제하기
-                </div>
-              </div>
-            </Overlay>
+            <PostMenuButton post={post} setType={setModalType}></PostMenuButton>
           ),
           delPost: (
-            <Overlay onClick={closeModal}>
-              <div
-                onClick={(e) => e.stopPropagation()}
-                className="rounded-2xl w-full h-[100px] p-4 m-4 flex bg-white self-end flex-col justify-evenly "
-              >
-                <div className="h-[30px] p-1">게시글을 삭제하시겠습니까?</div>
-                <div className=" flex divide-x w-full divide-gray-300 mt-2 mb-2">
-                  <button
-                    className="flex justify-center items-center flex-1"
-                    onClick={closeModal}
-                  >
-                    취소
-                  </button>
-                  <button
-                    type="button"
-                    className="flex justify-center items-center flex-1"
-                    onClick={postDelete}
-                  >
-                    확인
-                  </button>
-                </div>
-              </div>
-            </Overlay>
+            <Delete
+              onClose={closeModal}
+              onDelete={postDelete}
+              name="게시글"
+            ></Delete>
           ),
-          updateComment: selectComment && (
+          comment: (
+            <CommentMenuButton
+              onClose={closeModal}
+              setType={setModalType}
+            ></CommentMenuButton>
+          ),
+          updateComment: selectedComment && (
             <UpdateComment
-              selectComment={selectComment}
+              selectComment={selectedComment}
               closeModal={closeModal}
             />
           ),
           delComment: (
-            <Overlay onClick={closeModal}>
-              <div
-                onClick={(e) => e.stopPropagation()}
-                className="rounded-2xl w-full h-[100px] p-4 m-4 flex bg-white self-end flex-col justify-evenly "
-              >
-                <div className="h-[30px] p-1">댓글을 삭제하시겠습니까?</div>
-                <div className=" flex divide-x w-full divide-gray-300 mt-2 mb-2">
-                  <button
-                    className="flex justify-center items-center flex-1"
-                    onClick={closeModal}
-                  >
-                    취소
-                  </button>
-                  <button
-                    type="button"
-                    className="flex justify-center items-center flex-1"
-                    onClick={() => {
-                      commentDelete();
-                      closeModal();
-                    }}
-                  >
-                    확인
-                  </button>
-                </div>
-              </div>
-            </Overlay>
+            <Delete
+              onClose={closeModal}
+              onDelete={commentDelete}
+              name="댓글"
+            ></Delete>
           ),
         }[modalType]}
 
@@ -448,17 +334,14 @@ export default function ClubPost() {
             </div>
           </div>
           <div className="flex items-center">
-            <p className="text-sm">
-              댓글{" "}
-              {commentData?.data.length !== 0 ? commentData?.data.length : 0}개
-            </p>
+            <p className="text-sm">댓글 {commentData?.data?.length}개</p>
           </div>
         </div>
         <section className="mt-4 overflow-scroll">
           <ul>
-            {commentData?.data.map((item: any, idx: number) => {
+            {commentData?.data.map((item: any) => {
               return (
-                <li className="flex space-x-2 mb-3 relative" key={idx}>
+                <li className="flex space-x-2 mb-3 relative" key={item.id}>
                   <img
                     src={`${item.profileImg}/avatar`}
                     alt="유저 프로필"
@@ -478,7 +361,7 @@ export default function ClubPost() {
                       userData?.id !== item.userId ? "hidden" : ""
                     }`}
                     onClick={() => {
-                      setSelectId(item.id);
+                      setSelectedComment(item);
                       setInModal(true);
                       setModalType(ModalType.COMMENT);
                     }}
@@ -494,18 +377,10 @@ export default function ClubPost() {
       <BottomTabNavigator className="justify-between p-2 nav">
         <form onSubmit={handleSubmit(onSubmit)} ref={formRef}>
           <textarea
-            rows={1}
-            cols={20}
-            wrap="hard"
             required
             placeholder="댓글을 입력해주세요."
             className="bg-gray-100 py-4 px-3 rounded-lg flex-1 outline-none w-[300px] h-[48px] resize-none text-[16px] max-h-[48px]"
             {...register("comment")}
-            onChange={handleResizeHeight}
-            ref={(e) => {
-              ref(e);
-              commentRef.current = e;
-            }}
           />
         </form>
         <Button onClick={clickHandler}>전송</Button>
